Commit once per sent or deferred message
This makes a lot of sense anyway as the log entry and message deletion or
deferral should be treated as one atomic unit. It also has the benefit of
reducing the number of times we commit in the database, which can have a
huge impact on bulk sending speed.

For each sent message, we are now down to four database actions: SELECT the
message to send, INSERT the success log entry, DELETE the original message,
and COMMIT the whole result.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>

Conflicts:

	mailer/engine.py